requirements.txt
cycler==0.11.0
docopt-ng==0.7.2
fonttools==4.33.3
future==0.18.2
hal-cgp==0.1.1
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.2
mpmath==1.2.1
nolds==0.5.2
numpy==1.18.5
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
sympy==1.10.1
threadpoolctl==3.1.0
